backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
python-dotenv
sentence-transformers
chromadb

langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.3
langchain-text-splitters==0.2.0
langchain-google-genai
langchain-chroma
langchain-huggingface